Update virtual_person_id type to uint64.
Currently the population_offset and total_population in VirtualPersonPool are uint64, while virtual_person_id in VirtualPersonActivity is int64. This could cause an issue when the range covered by virtual person pools is out of range of int64.

This is a safe change per https://developers.google.com/protocol-buffers/docs/proto#updating